Zip the UCAS data matching file
