+++ .env.example
@@ -17,5 +17,6 @@ DSI_API_SECRET=xxxxxxxxxxxx-xxxxx-xxxxxxxx-xxxxxxxx
 SANDBOX=false
 UCAS_USERNAME=abcdef
 UCAS_PASSWORD=123123
+UCAS_ZIP_PASSWORD=8712634
 UCAS_UPLOAD_BASEURL=https://transfer.ucasenvironments.com/api/v1
 UCAS_UPLOAD_FOLDER=folder
