+++ .env.test
@@ -7,5 +7,6 @@ DSI_API_URL=https://test-api.signin.education.gov.uk
 DSI_API_SECRET=some-secret
 UCAS_USERNAME=abcdef
 UCAS_PASSWORD=123123
+UCAS_ZIP_PASSWORD=98126312
 UCAS_UPLOAD_BASEURL=https://transfer.ucasenvironments.com/api/v1
 UCAS_UPLOAD_FOLDER=685520099
