+++ docker-compose.azure.yml
@@ -26,6 +26,7 @@ services:
       - DSI_API_SECRET
       - UCAS_USERNAME
       - UCAS_PASSWORD
+      - UCAS_ZIP_PASSWORD
       - UCAS_UPLOAD_BASEURL
       - UCAS_UPLOAD_FOLDER
       - SANDBOX
